requirements.txt
aiofiles==23.2.1
beautifulsoup4==4.12.3
nextcord==2.6.0
playwright==1.44.0
python-dotenv==1.0.1